Allow alternative to linear srgb when xyb with ICC
For decoding a XYB-encoded image with ICC profile as original color
profile, allow configuring the decoder to decode to a color encoding of
your choosing, rather than the decoder always converting xyb to srgb.

Reason for this: the decoder is able to convert xyb to all supported
color encoding values (srgb, pq, hlg, linear, p3, ...). It does this
when it knows the color encoding, but not when an ICC profile is used
(since the libjxl decoder does not have color management system to parse
it, nor would all icc profiles be representable as a jxl color encoding)

If the decoder does not know the color encoding, it default to linear
srgb.

However, if you have a system, like the chrome browser, that can
determine form the ICC profile what the color encoding is, then it is
better to be able to let the decoder convert the xyb immediately to
that, to avoid needing two passes over the data.

Therefore this option is added to the API and implemented in the
decoder.