Check for AUTHORS file updates and re-enable PR checks.
Pull Request checks should only run on pull_request events, and not
needed to run on pull_request_review. This gets the feedback to the PR
as soon as possible. This was also causing issues in the PR with the
way we get the SHA values from the current environment.

This patch fixes the pull request workflow which was needed for the
AUTHORS check and adds this new check to update the AUTHORS file when
new contributors send patches. We now pull 10 commits from the HEAD of
the pull request (a PR normally has only 1 commit) and run the checks
on the PR commit.