New GitHub workflow to build and test several modes.
This new build_test.yml workflow builds and tests for release, debug,
asan, msan and release with scalar-only modes in the the main and
release branches.

This patch now also builds all these targets but only tests release in
the pull request, unless it is labeled with "CI:full" label which will
cause it to test all the modes as well.