Refactor decoding Jxl to JPEG out of decode.cc (#93)
This patch groups most of the JPEG reconstruction decoder logic out of
the JxlDecoder to a new class JxlToJpegDecoder that is included in the
JxlDecoder struct. This clears a little bit the JxlDecoder but more
importantly makes it easier to remove the dependency on jxl::jpeg::
functions with a flag at build time which will be added later.

(cherry picked from commit da7ff9a1956ea7749d804880bffd47ad6dc516a0)