Header only: rewinding API
Add proposed API for rewinding the decoder with the purpose of
efficiently skipping to a frame

The goal of this is as follows: if you are rendering a looping animated JXL
image, but do not keep all frames in memory, and need to render an
arbitrary frame in the animation again (which may be beyond the start,
and may be at any point in time, e.g. due to not rendering frames when
not looking at the animation), how can we decode that frame as
efficiently as possible again, given the encoded jxl input data:

Decoding every frame from the beginning is very inefficient if you only
need to display frame N, and the SkipFrames function helps with that.

Even then it may be doing too much work though, since a decoder reset to
the beginning doesn't know which patches / to-be-blended frames affect
the frame you are skipping to.

The rewind function allows the decoder to begin from the beginning and
keep state to remember which frames depend on which, to make that
scenario more efficient as well.