New JPEGXL_ENABLE_TRANSCODE_JPEG build flag to disable decoding to JPEG (#97)
JPEG reconstruction in the API is not needed if we are only interested
in decoding to pixels. Disabling support for decoding to JPEG in the API
saves about 95 kB in libjxl_dec.so in an arm build, and about 50 kB in a
size-optimized arm build.

(cherry picked from commit 5973b44bc36f2c4ca182978ec28f8503a9975985)