Use up all bytes of box to skip
In decode.cc, when parsing a box that is being skipped, "consume" its
bytes, rather than passing them back to the user: it is more efficient
for the user if the decoder indicates more bytes as consumed, and when
skipping a box there's no reason to require keeping them around. This
can increase performance depending on the scenario in which the input is
being passed.