Fix counterclockwise order for progressive
Before this fix, the groups in the boundary of the concentric squares
when using middleout progression didn't appear consistently in
counterclockwise nor counterclockwise order, but somewhat random
depending on the position of the center of the image in relation to the
center of group which contains the center. This was due to the fact that
for images with dimensions not a multiple of the group size, the center
was not computed correctly.
Now it works as intended: counterclockwise fills the boundary
of the next concentric square from a square on the left side of the
boundary.