Remove DC implementation
This removes the support for getting the DC image of frames as it was
currently implemented, it did not work several types of images, and to
be useful as a way to get downscaled image needs to be implemented
differently.

A currently available alternative is to use JxlDecoderFlushImage, but
at this time it's not possible to know when exactly it should be called
to get exactly dc only.

The functions are not removed as symbols, but they have no effect. The
DC feature already had no effect on several types of images and already
didn't guarantee that a DC would be returned.

Also removed usage of dc from the fuzzer