Move changelog to a CHANGELOG.md (#167)
The debian/changelog file is for the Debian package, which should list
changes made to the debian package, like changed flags or dependencies,
added patches to fix existing versions, etc. In our case, we normally
only just bump the version of the library and make no changes to the
package, so "Bump JPEG XL version to ..." is an appropriate message
there.

This patch moves all the available Changelog information to a new
CHANGELOG.md file following the guidelines from
https://keepachangelog.com/en/1.0.0/ . This will allow us to keep track
of notable changes when they land and make the release process easier.