Add reference getters to dec_frame
Add getters for which frame storage locations are referenced, and
referencable by, frames. This is for use for more optimal rewinding in
the API, but it is not implemented yet there.

Also added a test for a jxl image with patches to decode_test, to
prepare for when this feature is added to the API.