Store frame dependency graph on rewind
This stores the dependency graph between frames on rewind, so that
decoding of more frames can be avoided when skipping to a later frame
after rewind.

Improved the SkipFrameWithBlendingTest to have true dependencies between
frames that use a blending mode that depends on the pixels of both
frames, to test this feature, and verified that making
GetFrameDependencies return wrong output, breaks the test.