Support low memory if exif orientation is kept
In dec_frame, the low memory path was being disabled if the image has
exif orientation, even if the user wants to output as-is without
applying the orientation. This adds a new flag to indicate what is
intended, so the low memory path can be used if no orientation is
performed.

Also adds a test for orientation in decode_test (both with and without
keep_orientation).

Also fixes decoder rewind to keep the orientation setting around, since
settings are expected to stay after rewind.