Make GetSomeTestImage more varied
Make the randomly generated image from GetSomeTestImage more varied:
before, it always had a quarter circle of the same size in the top left
corner, and random noise elsewhere.

Now, it has a circle with random radius at random location, a rectangle
with random noise, and a random colored gradient as background.

I needed this to generate a jxl image with blending that has the
blending visible (multiple circles at random locations overlaid, which
was not possible to see if circle was always the same), but
this is worth committing because it improves the tests in general.