Merge multiple DoUndoXYBInPlace() paths.
Remove the template parameter from DoUndoXYBInPlace and replace the call
to Op::Transform with a virtual function call to a common base class
type. This stops inlining the call to Transform but allows the compiler
to not have multiple versions of DoUndoXYBInPlace().

In the x86 build with 4 SIMD targets, dec_reconstruct.o contribution to
libjxl_dec.so goes from 67 KB to 51 KB. In arm 32-bits with only one
target this reduces 2.7 KB.