Always define the fuzzer build targets
When JPEGXL_ENABLE_FUZZERS is off, we weren't defining any of the fuzzer
targets except in coverage builds (when JPEGXL_ENABLE_COVERAGE is on).
This patch instead defines the fuzzer targets as "djxl_fuzzer_runner"
when JPEGXL_ENABLE_FUZZERS is off but excluded from the "all" target.

There are no changes to the targets defined when JPEGXL_ENABLE_FUZZERS
is on, except that djxl_fuzzer depends on the decoder side only, helping
the fuzzer library with the coverage analysis by removing all the unused
encoder and extras library code.