Fuzz: Fix InvSqueeze when using channels with no pixels
The short-circuit for channels with no pixels was not updating the size
of the output channel, only updating the vshift/hshift. This caused the
channel sizes to get out of sync with the MetaApply ones. This patch
fixes that and also prints the squeeze transforms when debug is enabled
and only the default transform is used.