Use the correct CustomTransformData
There were two independent copies of CustomTransformData in CodecMetadata:
one directoy in CodecMetadata, the other in ImageMetadata (which is the
field m of CodecMetadata)

The one being saved in the file was the one from CodecMetadata, the one
used while decoding was instead the one from the ImageMetadata.

So it ignored the opsin inverse matrix from the jxl file, due to using
the other, default initialized, one for the actual decoding.

This removes the copy from ImageMetadata, so there's only one copy
left that's both saved and used.