Fixing the build in msan (#278)
msan needs a higher butteraugli score than non-msan

This is of course not ideal, the result should be the same.

PR 273 that surfaced this however did not introduce this discrepancy,
and rolling it back would not solve the root cause -- so 'fixing'
the problem with upping the limit. I have filed a bug about the
msan and non-msan not matching and we will discuss that later.