Fix YCbCrUpsampling handling of border.
When the image right or bottom border is not in the currently processed
group but the required padding around the group would go past the image
border in the next group, we would not mirror the right column/row for
YCbCr upsampling algorithm. This patch fixes the logic when mirroring
the last row / column.