Relax x0 % kPaddingXRound requirement in FilterPipeline
The FilterPipeline required that the x0 in all three input, output and
image rects be a multiple of kPaddingXRound. This has the issue that it
is 4 in ARM but 8 in x86, however Sigma is always an 8x8 downsampled
image.

This patch relaxes the requirement so that input, output and image rect
all have the same alignment % kPaddingXRound but it doesn't need to be
0, removing the logic that extends the filters ranges from
dec_reconstruct.

The EPF functions were receiving a image_x_mod_8 value as
image_rect.x0() % 8, which was always 0 in x86, but may be 4 in ARM.
This value now can be anything between 0 and 7 in both architectures
since it depends on the value of x0.

This patch overall reduces a tiny bit the total range of [x0, x1) where
we run intermediate filters since the FilterPipeline now handles the
values of image_rect.x0() that are not a multiple of kPaddingXRound.

This removes the rect parameter from ApplyFiltersRow() since it is
already passed on initialization and it must be the same every time.