Update cmake project() name (#321)
The project() cmake command defines the ${project_name}_VERSION variable
with the value passed in the VERSION parameter to this command. We are
currently not setting the VERSION parameter so this was defining
JPEGXL_VERSION as the empty string. Independent to this we where using
the JPEGXL_VERSION cmake variable to let integrators define a project
version to include in the tools (like the package version) which was not
working since cmake would override its value.

This patch renames the project() name to LIBJXL to workaround this
issue. We should eventually set up the version numbers in the project()
as well.