Fix name mismatch "associated" and "premultiplied"
Fix name mismatch in the API. These are corresponding fields and
it was a historical error that they appeared independently with
a different name.

Also add to the comments that they match that of the main alpha
extra channel.

This renames a field in the API, but I believe that because the API
did not yet support outputting extra channels, usage of this field
should be extremely rare.