Fix Threads::Threads definition when JPEGXL_STATIC=1
This sets Threads::Threads to use -Wl,--whole-archive for -lpthread when
compiling with JPEGXL_STATIC=1 since otherwise weak symbols in libstdc++
would not be resolved to pthread's library. This is more of a workaround
for compiling static binaries. The fix modifies the link libraries
properties of Threads::Threads.

Also fixes the -lrt flag in benchmark::benchmark making the benchmark
also compile in static mode when JPEGXL_STATIC=1. There are no changes
to the build when JPEGXL_STATIC=0.

Fixes #332.