Fix pixbufloader-jxl and add a test (#336)
The pixbufloader-jxl was not destroying the decoder object after
decoding and therefore leaking memory. This patch fixes that by adding
the missing call to `JxlDecoderDestroy`.

A new test will try to load a .jxl file from disk using the
pixbufloader-jxl plugin. This would detect leaks when running in asan
mode.

A drive by fix, the pixbufloader doesn't need the full library, only the
decoder side.

Fixes #299