Set JPEGXL_VERSION macro from git hash at build time.
This sets the JPEGXL_VERSION macro used in the tools from the short git
hash at build time, updating it only when needed to avoid rebuilding
every time. Unfortunately, ninja checks for the header timestamp before
running the target that generates it so when only updating the hash, the
tool_version.cc.o file is not regenerated. This doesn't affect make, or
when running `ci.sh release` or similar commands.

GitHub/GitLab builders use the hash of the built reference, so this code
doesn't run in the workflow/pipelines anyway.

Tested with:
  - `./ci.sh release`
  - `./ci.sh release -DJPEGXL_VERSION="foo bar"`
  - `mkdir build; cd build; cmake ..`

Test rebuilt after amending a commit.