Fuzz: Fix Rect rounding error decoding extra channels.
When using an upsampling factor > 1 for extra channels with an odd size
(or not a multiple of the upsampling factor) we need to round up the
Rect copied in ModularImageToDecodedRect().

This fixes an msan bug where some pixels near the right or bottom edge
can be uninitialized.

Added more msan checks to the extra channels to fail faster on this type
of errors.