Print stack trace on ASSERT() when building in oss-fuzz (#364)
JXL_ASSERT() and similar functions use the sanitizer's own library to
print a stack trace when an assert triggers. This was only enabled when
MEMORY_SANITIZER, ADDRESS_SANITIZER or THREAD_SANITIZER was defined,
which is not the case in the ossfuzz builds. Now this is controlled by
the compile flags like -fsanitize=address.