Initialize all X and Y padding on extra channels in msan and oss-fuzz.
Instead of initializing only the padding in X for all the rows in the
extra channels this initializes the whole row for the rows that are
extra padding at the bottom of the channel.

This shouldn't be needed but the extra channels are processed up to the
upsampled padded size.

Using JXL_MEMORY_SANITIZER enables this in the oss-fuzz builds as well.