Fuzz: Check for out of bounds access in squeeze channels
squeeze in_place transforms use the residual channels from right after
the squeeze transform range, but that might be out of bounds if the
endc channel is near the end of the list and there's some other
programming error. This patch checks for the channel number before
accessing it with an assert to fail earlier and help the fuzzer detect
these cases.