Fix division through zero issue in AdaptToXYZD50
The division through zero was not the division through wy, but a value
that happened to be 0 after the bradford matrix multiplication.
Requiring the values to be in range 0-1, similar to skcms, should fix
this.
Removed the std::finite checks since requiring wy to be > 0 already
fixes this.
Also did this in the similar PrimariesToXYZD50.
Added TODO for also checking the r,g,b primaries like this: the ACES
P0 color space supports out of range values here.