Fuzz: Return error if Gaborish weights lead to invalid kernel. (#371)
Spec says that the coefficients of the kernel used in Gaborish are
normalized to sum 1, but if the sum is already near 0 this is undefined.

This patch catches that case and returns a JXL_FAILURE instead of
creating NaNs or infs in the output.
(cherry picked from commit e44964d3c04cd6a6545a265224f04de7d525c3f1)
(cherry picked from PR #371)