Fix modular nb_meta_channels counter (#373)
Fuzzing hasn't found a problem case yet afaik, but in the palette transform there was a similar problem as in #370 : the nb_meta_channels counter was not properly decremented in the inverse transform.

Also now checking for doing transforms on an invalid mix of meta/nonmeta channels in CheckEqualChannels(), so it will also cause RCT to fail when tried to be applied to a mix of meta and nonmeta channels.
(cherry picked from commit 3516cd58cee882e6905139b303873606b26720f1)
(cherry picked from PR #373)