Fuzz: return early when there is no modular data to decode (#380)
When a modular dc or ac group is empty, a 0-channel image gets decoded, and in the old flow that just meant that all of those 0 channels would get copied to the modular full_image.

In the new flow however (where full_image is avoided), it created trouble because a 0-channel image would still cause a call to ModularImageToDecodedRect(), which would then try to access non-existing channels.

Returning early when a 0-channel group image is about to get decoded solves that.
Still adding an assert to check for access to non-existing channels.

Also don't do the no-full-image thing when not all modular channels have the same shift. That avoids the potential problem where some channels are encoded in the DC groups and others in the AC groups, which would cause trouble since ModularImageToDecodedRect() assumes all channels are available. This is a temporary fix; a better but more complicated fix would be to update ModularImageToDecodedRect() to work on the subset of channels that was just decoded.