Build .deb packages inside a Docker container
The "ubuntu-18.04" GitHub environment is not exactly the same as
upstream's 18.04 from Ubuntu. In particular, the built libjxl .deb
package gets a dependency on libgcc-s1 which is not available in
Ubuntu 18.04.

This patch builds the Debian packages using upstream's Docker container
instead. This also allows to build for other Debian-based environments
not otherwise available in GitHub, such as debian:buster and
debian:bullseye which are now added to the list.