Move the DecoderHints out of the libjxl library.
The DecoderHints are hints not for the JXL decoder, but for decoders of
all the other formats supported by the extras libraries. These "hints"
are only ever used for specifying the "color_space" (Description of
ColorEncoding) or the path to an ICC profile blob. This is useful when
loading images from format that don't specify the color space (like PNM)
and in cases where the color space is just not included in a given file.

The hints were part of the CodecInOut but were really only used by the
decoders in extras/ as an input together with the encoded image bytes.

This patch moves the DecoderHints to the extras/ library and renames it
to ColorHints since this is only used for color encoding.

One small side-effect of this refactor is that now the GIF and PSD
decoders supports parsing the color hints if passed.