Move ColorEncoding ParseDescription() to extras
ParseDescription() converts a textual description of the ColorEncoding
to the internal object representation. This is only used by the tools to
let the user pass a color encoding to formats like .pnm that don't have
a color encoding defined in the format.

This patch moves the ParseDescription() function to extras and uses the
external JxlColorEncoding type instead of the internal one when parsing
the description in preparation for tools to be able to pass this object
through the public API.

The matching Description() function is still in the libjxl since it is
used in one place in the decoder when creating an ICC from a JPEG XL
color encoding object. The serialized description is attached as a tag
in the ICC in that case. While this tag doesn't need to conform to any
format it is convenient to have a description of the encoded ICC. The
description encoding side is not too much code anyway, but the
description parsing was actually unused from within the library.