Fix importing external JxlColorEncoding with custom cases.
Primaries and WhitePoint custom values should only be used when the enum
values are the `*_CUSTOM` ones, otherwise those values can be undefined.

This patch fixes handling those cases and cases where the enum values
are unknown (invalid) to help with forward ABI compatibility if new enum
values are added in the future.