Header only: box API design
This MR proposes API for getting metadata boxes in the decoder.

It is integrated with the event based system the decoder uses,
using the JXL_DEC_BOX event to indicate a new box.

Boxes are output to a byte buffer owned by the user, similar to the
recompressed JPEG output.

Boxes can be gotten in an uncompressed mode or in a raw mode. For
non-brob boxes, these modes are the same. For brob boxes, the standard
mode automaticlaly decompresses them, the raw mode returns them as-is.