Auto resample for high distance (#534)
By default, let cjxl choose resampling=2 if requested distance >= 20

This also increases the allowed --distance given to cjxl to go up to 25,
previously it only was allowed up to 15.

For these bit rates, using the 2x2 resampling looks better on most
images.

The distance itself is lowered to 6 (and slightly more for higher
original distance), which roughly ends up at similar bit rate. This
depends on the image but is relatively close overall.

While there are lower distances where rescaling works better as well, it
depends on the image, 20 is a conservative choice.

This only implements it for the cjxl binary tool for now, not for the encoder
API (which doesn't yet support the resampling option; to be added soon)
nor for the benchmark.

To force using no resampling anyway, you can still use resampling=1.