Remove JPEG's DecodeTarget parameter.
DecodeTarget was telling DecodeImageJPG whether we should load the pixel
or the JPEG coefficients. To load the JPEG coefficients we were actually
using the library internal API. Loading the coefficients is used for
pixel-lossless (but not byte-lossless) encoding of JPEG files. This is
not currently supported in the public API, but it would certainly be
a different workflow than loading the pixels (either we pass an image
with coefficient values or we pass the JPEG file like in lossless
recompression).

This patch splits out this workflow from DecodeImageJPG and
EncodeImageJPG so we can start migrating those functions to use the
external types.