Fix building with LLVM LLD
In 63a481a, linker flag "-Wl,--exclude-libs=ALL" is added.
However, LLVM LLD does not support it.
Building with LLVM toolchain would result in failure.

This commit check whether the linker support this flag.
If it does not, the flag will not be added.
A not-so-pretty method with CheckCSourceCompiles is used,
because CMake 3.10 does not support
CheckLinkerFlag or CMAKE_REQUIRED_LINK_OPTIONS.