Implement cmake's BUILD_SHARED_LIBS setting. (#575)
The current behavior was to build both shared and static libraries since
tests and some tools use internal symbols not available in the shared
libraries. We support passing `-DJPEGXL_STATIC=ON` to build all the
binaries as static binaries, which causes them to link against the
static version of the libraries but also link as a static binaries with
all the problems that such thing brings.

The new `-DBUILD_SHARED_LIBS=OFF` setting, which is the standard name
in cmake for this behavior, would make jxl, jxl_dec and jxl_threads all
be aliased to the -static targets so tools and plugins link against the
static version of those.