Remove Fields::Name() strings from non-debug builds.
The const string returned by Name() ends up in the library and it is
only used in debug builds when Fields tracing is enabled. This patch
removes the Name() virtual method and replaces it with an empty string
in non-debug build so these strings don't end up in the library.