Move the check for JPG files earlier.
jpeg::DecodeImageJPG() can return an error after modifying the
CodecInOut on invalid JPEG files (or non-JPEG files). This causes codecs
running after the call to DecodeImageJPGCoefficients() to fail to decode
unless -j is passed to cjxl. This solves the issue at least for valid
JPEG files and files that not start with a JPEG header. File that look
like a JPEG might trigger an error but those shouldn't be parsed by any
other codec on the chain.

This is not an issue when calling jpeg::DecodeImageJPG() from the API
because the CodecInOut is not reused afterwards.