Load from other codecs to an interleaved image format.
This patch defines a new set of classes like PackedImage that represent
the image concepts as define in the external C API. These classes
represent a jxl decoded image that will be used as the output from the
the decoder API and input to the jxl encoder API.

Migrated most codec parsers (all excepts PSD) to use this format to load
still and animation images without extra channels (besides alpha).

Manually tested .gif and .exr files. .png and .jpg are extensively
tested by the unittests.