Remove unused jxl::Divider() (#637)
The `class Divider` implemented a unsigned 32-bit division that
accelerated dividing by the same divisor multiple times. This could be
of use in a libjxl_threads implementation (and that's why it was in
data_parallel.h) but this would be out of the scope of libjxl and it
hasn't been used in a long a time.