Allow setting more boxes after decoder finished
The box based container format doesn't indicate which box is the last
(having an unbounded size last box is possible but not required), so the
decoder cannot know for sure when the file finished.

If there would be metadata boxes after codestream boxes, the decoder
could return JXL_DEC_SUCCESS when it decoded all codestream boxes and
the user input happened to end exactly at the last codestream box
boundary.

The API did not allow calling JxlDecoderProcessInput again after it
returned success, so it would be impossible to decode those further
metadata boxes.

This MR changes that by allowing to call JxlDecoderSetInput and then
process the further input again.

It will still give errors when calling JxlDecoderProcessInput without
setting any new input, and also when it's not the container format, or a
codestream box is encountered even though all codestream boxes were already
decoded.