fix allocation not being done before use in some cases (#655)
In some cases (progressive rendering with a flush at an unfortunate moment), the allocations done in InitForAC() were not yet done when they are already expected to be done to be able to render partial images, causing segfault. I think this could happen when DC is available but ACGlobal not yet, or something like that. This fixes that.
(cherry picked from commit 4049834fa68bb54480932d9c165edeecd198c778)
(cherry picked from PR #655)