Fix issue with partial codestream boxes (#664)
Fixes issue discovered in https://github.com/libjxl/libjxl/pull/651/:

It was saving previous box data to the codestream buffer when a next
partial codestream box is encountered. However, if there were other
unrelated boxes in between, it already updated position pointers and no
longer included the previous box data.

Now it instead saves the codestream data to the internal buffer
immediately if it's a partial codestream, rather than waiting for the
next one. It has to do this anyway to concatenate the separate buffers,
so the memory usage is about the same.

Also removed variable box_begin that was never read from.