Use PRIuS/PRIdS macros to printf size_t/ssize_t.
PRIuS is defined to either "zu" or "Iu" depending on the platform
allowing to print size_t values in MSYS2. This macro is meant to be used
similarly to PRIu32, so that the "%" is not part of the macro allowing
expressions like `"%3" PRIuS` instead of `"%3zu"`. Similarly, `PRIdS` is
used for ssize_t.

This patch replaces all the occurrences and adds a test to bash_test.sh.

Fixes #718.