Fix whitepoint, primaries & tf CICP documentation
Those enum values are only a subset of the full set of possible CICP
values, so the documentation was not entirily correct.

Make clear that it's only a subset, and also that you can always
read the values from the floating point xy primaries and whitepoint
fields, regardless of the enum value.