Fix std::max with zero issue (#745)
I accidentally found few lines in the code where `std::max` is used wrong: `std::max(0, /* size_t */ something)`. Obviously, the purpose of such an expression was to return zero instead of a negative value.

But both `x` and `xborder` are unsigned integers and subtraction of them is also unsigned. So we get underflow here.

Updated AUTHORS