Fix out of bounds copy in LoadBorders() (#775)
When processing groups out of order with an image where the last group
size  is smaller than the needed border it was possible to attempt to
load the right or bottom border from the next group when already
processing the rightmost or bottom group respectively. This situation
was causing an out-of-bounds copy on saved Image3F buffer in release
mode (or hitting a JXL_DASSERT in debug mode).

The order in which the groups are processed depends on many factors,
including the order in which the threads are scheduled when using
multiple threads, and potentially the order of the groups in the file
(not checked).

Added a test to force the out-of-order situation in a simulated parallel
runner that forces a random order on the tasks. The new test triggers
the assert in debug mode, which is now fixed.

(cherry picked from commit e6497057899269bb40f54a26729826a55d857fd7)
(cherry picked from PR #775)

Adapted the cherry-pick to use <random> in the test instead of
"lib/jxl/random.h" since the latter is not available in v0.6.x branch.