Migrate to AButler/upload-release-assets action.
svenstaro/upload-release-action is triggering connection issues when
uploading artifacts to releases and there's no fix for it. Migrate to
a different action for uploading artifacts to releases.

Added `fail-fast: false` to increase the likelihood that other workflows
in the matrix upload their artifacts when one fails so we can retry
less jobs.