Add modular setting to encoder
Add setting to the encoder to force modular or vardct mode, or use the
default (where in future implementations, encoder is allowed to e.g. use
heuristics on pixel content to choose this)

Also lets the SetDistance option set the quality_pair setting for
modular too. As dicussed with Jon, it's ok to translate distance to
a quality_pair value and expose only one setting in the API, rather than
expose the distance setting for vardct, and the quality pair setting for
modular, separately. The conversion formulas come from cjxl.cc and can
be tweaked if necessary.

Also alters the documentation of the SetLossless setting to specify what
this setting actually does as implemented, and do not be as strict in the
documentation which settings exactly this overrides since this can be
future implementation dependent.