Move printf macros PRIuS/PRIdS to printf_macros.h (#810)
PRIuS/PRIdS are the standard names for these macros that should be
defined by the system headers. We have an ifndef guard against double
definition, however it is possible for our header to be included before
a system header (in particular gtest) so this fails with a double
definition of the macro.

Move these definitions to a new header to avoid this issue.
This patch also adds a missing include header in
fake_parallel_runner_testonly.h.